# ⚡️ Speed up function `_kvform_pairs_to_dict` by 1,393%