@aseembits93 2025-08-28
Hi @qued ! Thanks for your time. I was wondering if I need to update the CHANGELOG and __version__ here? I'm asking because there are 2 more PRs like this and I don't know in what order the PRs would be merged. Thanks in advance!