@qued 2025-08-28
Can you explain what "Generated Regression Tests and Runtime" is? I don't see tests added to the repo, so I assume this is a test suite that's created and run ephemerally?

Also with one test that I'm interested in, `test_large_list_preserves_deepcopy`, I see where the deep structure of the output is being modified, but then I don't see where we check that the original artifact is unmodified. This would be my main worry: The writer of the original code presumably used `deepcopy` because they want to ensure the new artifact is not re-using objects in the passed artifact, so I'd want to ensure the new code is following suit.