@aseembits93 2025-08-28
@qued "Generated Regression Tests and Runtime" refer to AI generated tests which were used in addition to the existing tests in the codebase to enhance the testing in case we don't have any or very few tests directly calling the function under consideration. These tests ensure new changes do not break or negatively affect the existing functionality of an application. I was closely testing this function after your comment and found that the codeflash suggested change actually mutates the input which is not desirable. In the past we have found optimizations which skip deepcopy but maintain the immutability of the input. Thank you for the catch, I will close this PR.